vpc/requirements.txt
-e .
aws_cdk.aws-ec2==1.102.0
